refactor: Update imports for AccountProvider in Views components
